refactoring: extract all package-json related code to a new module
