+++ src/consumer/consumer.js
@@ -20,9 +20,7 @@ import {
   COMPONENT_ORIGINS,
   BIT_VERSION,
   NODE_PATH_SEPARATOR,
-  LATEST_BIT_VERSION,
-  CFG_REGISTRY_DOMAIN_PREFIX,
-  DEFAULT_REGISTRY_DOMAIN_PREFIX
+  LATEST_BIT_VERSION
 } from '../constants';
 import { Scope, ComponentWithDependencies } from '../scope';
 import migratonManifest from './migrations/consumer-migrator-manifest';
@@ -35,7 +33,7 @@ import ComponentMap from './bit-map/component-map';
 import { MissingBitMapComponent } from './bit-map/exceptions';
 import logger from '../logger/logger';
 import DirStructure from './dir-structure/dir-structure';
-import { getLatestVersionNumber, filterAsync, pathNormalizeToLinux, pathRelative } from '../utils';
+import { getLatestVersionNumber, filterAsync, pathNormalizeToLinux } from '../utils';
 import loadDependenciesForComponent from './component/dependencies-resolver';
 import { Version, Component as ModelComponent } from '../scope/models';
 import MissingFilesFromComponent from './component/exceptions/missing-files-from-component';
@@ -43,7 +41,7 @@ import ComponentNotFoundInPath from './component/exceptions/component-not-found-
 import npmClient from '../npm-client';
 import { RemovedLocalObjects } from '../scope/component-remove';
 import { linkComponents } from '../links';
-import { getSync } from '../api/consumer/lib/global-config';
+import * as packageJson from './component/package-json';
 
 export type ConsumerProps = {
   projectPath: string,
@@ -605,7 +603,7 @@ export default class Consumer {
     }
     await bitMap.write();
     if (installNpmPackages) await this.installNpmPackages(componentsWithDependencies, verbose);
-    await this.addComponentsToRootPackageJson(writtenComponents, bitMap);
+    await packageJson.addComponentsToRoot(this, writtenComponents, bitMap);
 
     return linkComponents(
       componentsWithDependencies,
@@ -615,83 +613,6 @@ export default class Consumer {
       this,
       createNpmLinkFiles
     );
-  }
-
-  static getRegistryPrefix() {
-    return getSync(CFG_REGISTRY_DOMAIN_PREFIX) || DEFAULT_REGISTRY_DOMAIN_PREFIX;
-  }
-
-  async addComponentsToRootPackageJson(components: Component[], bitMap: BitMap) {
-    const importedComponents = await filterAsync(components, (component: Component) => {
-      const componentMap = component.getComponentMap(bitMap);
-      return componentMap.origin === COMPONENT_ORIGINS.IMPORTED;
-    });
-    if (!importedComponents || !importedComponents.length) return;
-
-    const driver = await this.driver.getDriver(false);
-    const PackageJson = driver.PackageJson;
-    const componentsToAdd = R.fromPairs(
-      importedComponents.map((component) => {
-        const locationRelativeToConsumer = this.getPathRelativeToConsumer(component.writtenPath);
-        const locationAsUnixFormat = `./${pathNormalizeToLinux(locationRelativeToConsumer)}`;
-        return [component.id.toStringWithoutVersion(), locationAsUnixFormat];
-      })
-    );
-    const registryPrefix = Consumer.getRegistryPrefix();
-    await PackageJson.addComponentsIntoExistingPackageJson(this.getPath(), componentsToAdd, registryPrefix);
-  }
-
-  /**
-   * Only imported components should be saved with relative path in package.json
-   * If a component is nested or imported as a package dependency, it should be saved with the version
-   */
-  getPackageDependencyValue(
-    dependencyId: BitId,
-    parentComponentMap: ComponentMap,
-    dependencyComponentMap?: ComponentMap
-  ) {
-    if (!dependencyComponentMap || dependencyComponentMap.origin === COMPONENT_ORIGINS.NESTED) {
-      return dependencyId.version;
-    }
-    const dependencyRootDir = dependencyComponentMap.rootDir;
-    const rootDirRelative = pathRelative(parentComponentMap.rootDir, dependencyRootDir);
-    return rootDirRelative.startsWith('.') ? rootDirRelative : `./${rootDirRelative}`;
-  }
-
-  async getPackageDependency(dependencyId: BitId, parentId: BitId) {
-    const bitMap = await this.getBitMap();
-    const dependencyComponentMap = bitMap.getComponent(dependencyId);
-    const parentComponentMap = bitMap.getComponent(parentId);
-    return this.getPackageDependencyValue(dependencyId, parentComponentMap, dependencyComponentMap);
-  }
-
-  async changeDependenciesToRelativeSyntaxInPackageJson(components: Component[], dependencies: Component[]) {
-    const bitMap = await this.getBitMap();
-    const dependenciesIds = dependencies.map(dependency => dependency.id.toStringWithoutVersion());
-    const driver = await this.driver.getDriver(false);
-    const PackageJson = driver.PackageJson;
-    const updateComponent = async (component) => {
-      const componentMap = component.getComponentMap(bitMap);
-      let packageJson;
-      try {
-        packageJson = await PackageJson.load(componentMap.rootDir);
-      } catch (e) {
-        return Promise.resolve(); // package.json doesn't exist, that's fine, no need to update anything
-      }
-      const packages = component.dependencies.map((dependency) => {
-        const dependencyId = dependency.id.toStringWithoutVersion();
-        if (dependenciesIds.includes(dependencyId)) {
-          const dependencyComponent = dependencies.find(d => d.id.toStringWithoutVersion() === dependencyId);
-          const dependencyComponentMap = dependencyComponent.getComponentMap(bitMap);
-          const dependencyLocation = this.getPackageDependencyValue(dependencyId, componentMap, dependencyComponentMap);
-          return [dependencyId, dependencyLocation];
-        }
-        return [];
-      });
-      packageJson.setDependencies(packageJson.packageDependencies, R.fromPairs(packages), Consumer.getRegistryPrefix());
-      return packageJson.write({ override: true });
-    };
-    return Promise.all(components.map(component => updateComponent(component)));
   }
 
   moveExistingComponent(bitMap: BitMap, component: Component, oldPath: string, newPath: string) {
